% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codesccgdiff.R
\name{codesccg.diff}
\alias{codesccg.diff}
\title{find divergent codes between child and the caregivers ignoring personal letter of code scheme}
\usage{
codesccg.diff(data, code, code_cg)
}
\arguments{
\item{data}{data frame containing both children's and caregivers' codes}

\item{code}{character string indicating variable name of children's code}

\item{code_cg}{character string indicating variable name of caregivers' code}
}
\value{
character vector with children's codes only (!) that differ from their caregivers' codes
}
\description{
filters codes that do not match between the child and the caregivers ingnoring their personal letter in the code and returns the missmatching codes of the child only. The corresponding missmatiching codes of the caregivers are NOT returned.
}
