% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitivevariablerename.R
\name{cognitivevariables.rename}
\alias{cognitivevariables.rename}
\title{renaming cognitive data}
\usage{
cognitivevariables.rename(
  data,
  time = c("baseline", "follow-up 1", "follow-up 2"),
  save,
  file
)
}
\arguments{
\item{data}{file with cognitive data with column names to be renamed}

\item{time}{character indicating assessment occasion. Must be one of "baseline", "follow-up 1", or "follow-up 2"}

\item{save}{logical. If TRUE returned data set will be saved as .csv}

\item{file}{optional. absolute path including file name and appendix .csv in "" where returned data set should be stored}
}
\value{
data set with cognitive data with renamed column names
}
\description{
renaming variable names of cognitive data sets
}
