% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataconvert.R
\name{metadata.convert}
\alias{metadata.convert}
\title{convert metadata}
\usage{
metadata.convert(
  metadata,
  data,
  subject = c("Caregivers", "Children", "HeadTeachers", "NumeracyLiteracy", "Students",
    "Teachers", "TwinCaregivers"),
  country = c("Ghana", "Tanzania", "Uganda", "Haiti"),
  time = c("baseline", "follow-up 1", "follow-up 2"),
  save,
  path
)
}
\arguments{
\item{metadata}{data set containing exported meta data to be renamed}

\item{data}{data set containing exported assessment data}

\item{subject}{data set with exported variable names of data sets and how to be renamed}

\item{country}{character specifying the country participant codes were assessed. Must be one of "Ghana", "Tanzania", "Uganda", or "Haiti".}

\item{time}{character indicating of assessment period. Must be one of "baseline", "follow-up 1", or "follow-up 2".}

\item{save}{logical. If TRUE returned data set will be saved as .csv}

\item{path}{optional. absolute path where returned data set should be stored}
}
\value{
data set with renamed meta data
}
\description{
convert meta data to useful meta data set for renaming variable names in assessment data sets (not used for data aggregation!)
}
