% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadatadataimport.R
\name{metadata.import}
\alias{metadata.import}
\title{import data}
\usage{
metadata.import(path_data, path_metadata)
}
\arguments{
\item{path_data}{name of the file with assessment data, can include absolute path}

\item{path_metadata}{optional file name with absolute path with meta data - not for data aggregation!}
}
\value{
one or two data frames of imported data
}
\description{
import of data set with raw unprocessed assessment data (only .csv with , as separator) and an optional meta data set (.csv with ; as separator or .xls) of the respective survey
}
