% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcdatafiltering.R
\name{qcdata.filtering}
\alias{qcdata.filtering}
\title{quality control that filters participant codes not matching required coding scheme}
\usage{
qcdata.filtering(
  data,
  level2,
  code,
  code_cg,
  school,
  person = c("Teacher", "Child", "Mother", "Father", "Caregivers"),
  country = c("Ghana", "Tanzania", "Uganda", "Haiti"),
  save,
  path
)
}
\arguments{
\item{data}{data frame containing variables of participants' codes for quality control}

\item{level2}{logical to indicate data frame holds level 2 data. If TRUE school code instead of participant code will be quality controlled. If TRUE no @param code needed.}

\item{code}{character of variable in data frame that hold participants codes - note for numeracy & literacy data: pnl_code & snl_code to to be converted in _code as the following STG_data_renamed$BA_code <- ifelse(STG_data_renamed$BA_pnl_code == "", STG_data_renamed$BA_snl_code, STG_data_renamed$BA_pnl_code) before hand!}

\item{code_cg}{character of variable in data frame holding caregivers code - only for twin combinde data set. Has to be specified when @param person = Caregivers}

\item{school}{character of variable in data frame holding school code}

\item{person}{character specifying the person participants code belongs to. Must be one of "Teacher", "Child", "Mother", or "Caregivers".}

\item{country}{character specifying the country participant codes were assessed. Must be one of "Ghana", "Tanzania", "Uganda", or "Haiti".}

\item{save}{logical. If TRUE returned data frame will be saved as .csv with ; separation}

\item{path}{absolute path where returned data frame should be stored.}
}
\value{
data frame of participant codes not matching required participant coding scheme with all variables plus added "pattern" column holding information on discrepancy between participant code scheme and required scheme
}
\description{
this function filters cases (rows) that do not follow the required coding scheme set by the research team of AE17
}
