% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcdataimport.R
\name{qcdata.import}
\alias{qcdata.import}
\title{import data set for quality control of codes}
\usage{
qcdata.import(file_path)
}
\arguments{
\item{file_path}{name of the file with data to be quality controlled, can include absolute path}
}
\value{
data frame of imported data
}
\description{
import data set with data and code to be quality controlled. Only .csv are valid to be imported.
}
