% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcphasesfilteringBAFU1FU2.R
\name{qcphasesBAFU1FU2.filtering}
\alias{qcphasesBAFU1FU2.filtering}
\title{quality control that filters baseline participant codes not matching participation code & demographic info between assessment phases (BA, FU1)}
\usage{
qcphasesBAFU1FU2.filtering(data, person = c("Teacher", "Child"), save, file)
}
\arguments{
\item{data}{data frame}

\item{person}{character specifying the person participants code belongs to. Must be one of "Teacher" or "Child"}

\item{save}{logical. If TRUE returned data frame will be saved as .csv with ; separation}

\item{file}{optional. absolute path including file name and appendix .csv in "" where returned data set should be stored}
}
\value{
data frame of participant's unique UUID not not matching participation code & demographic info between assessment phases (Baseline, Follow Up 1, Follow Up 2)
}
\description{
this function filters cases (rows, baseline participation codes) that do not match participation code & demographic info between assessment phases
}
